+++ .php-cs-fixer.dist.php
@@ -17,6 +17,7 @@ return $config->setRules([
         '@PhpCsFixer' => true,
         // Override PhpCsFixer rules.
         'concat_space' => false,
+        'phpdoc_to_comment' => ['ignored_tags' => ['var']]
     ])
     ->setFinder($finder)
 ;