+++ src/Exception/ContainerEntryNotFoundException.php
@@ -1,0 +1,11 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Normalizator\Exception;
+
+use Psr\Container\NotFoundExceptionInterface;
+
+class ContainerEntryNotFoundException extends \Exception implements NotFoundExceptionInterface
+{
+}