+++ tests/Integration/Console/Command/FixCommandTest.php
@@ -4,26 +4,10 @@ declare(strict_types=1);
 
 namespace Normalizator\Tests\Integration\Console\Command;
 
-use Normalizator\Cache\Cache;
-use Normalizator\ConfigurationResolver;
 use Normalizator\Console\Application;
 use Normalizator\Console\Command\FixCommand;
-use Normalizator\EventDispatcher\Event\NormalizationEvent;
-use Normalizator\EventDispatcher\EventDispatcher;
-use Normalizator\EventDispatcher\Listener\NormalizationListener;
-use Normalizator\EventDispatcher\ListenerProvider;
-use Normalizator\FilterFactory;
 use Normalizator\Finder\File;
-use Normalizator\Finder\Finder;
-use Normalizator\NormalizationFactory;
-use Normalizator\Normalizator;
 use Normalizator\Tests\NormalizatorTestCase;
-use Normalizator\Util\EolDiscovery;
-use Normalizator\Util\FilenameResolver;
-use Normalizator\Util\GitDiscovery;
-use Normalizator\Util\Logger;
-use Normalizator\Util\Slugify;
-use Normalizator\Util\Timer;
 use Symfony\Component\Console\Tester\CommandTester;
 
 /**
@@ -35,40 +19,12 @@ class FixCommandTest extends NormalizatorTestCase
 {
     public function createApplication(): Application
     {
-        $finder = new Finder();
-        $gitDiscovery = new GitDiscovery();
-        $eolDiscovery = new EolDiscovery($gitDiscovery);
-        $slugify = new Slugify();
-        $filenameResolver = new FilenameResolver();
-        $cache = new Cache();
-        $filterFactory = new FilterFactory($finder, $cache, $gitDiscovery);
-
-        $logger = new Logger();
-        $normalizationListener = new NormalizationListener($logger);
-        $listenerProvider = new ListenerProvider();
-        $listenerProvider->addListener(NormalizationEvent::class, $normalizationListener);
-        $eventDispatcher = new EventDispatcher($listenerProvider);
-
-        $normalizationFactory = new NormalizationFactory(
-            $finder,
-            $slugify,
-            $eolDiscovery,
-            $gitDiscovery,
-            $filterFactory,
-            $eventDispatcher,
-        );
-        $normalizator = new Normalizator($normalizationFactory, $filenameResolver, $eventDispatcher, $logger);
-        $configurationResolver = new ConfigurationResolver($eolDiscovery);
+        /** @var FixCommand */
+        $fixCommand = $this->container->get(FixCommand::class);
 
         $application = new Application();
-        $timer = new Timer();
-        $application->add(new FixCommand(
-            $configurationResolver,
-            $finder,
-            $normalizator,
-            $timer,
-            $logger
-        ));
+
+        $application->add($fixCommand);
 
         return $application;
     }